+++ typescript/cli/src/commands/options.ts
@@ -6,6 +6,8 @@ import { LogFormat, LogLevel } from '@hyperlane-xyz/utils';
 
 import { ENV } from '../utils/env.js';
 
+import { ChainTypes } from './types.js';
+
 /* Global options */
 
 export const logFormatCommandOption: Options = {
@@ -105,6 +107,20 @@ export const agentConfigCommandOption = (
   } file path for the agent configuration`,
   default: defaultPath,
 });
+
+export const chainTargetsCommandOption: Options = {
+  type: 'string',
+  description: 'Comma-separated list of chain names',
+  alias: 'c',
+  // TODO choices: ... can we configure this so that it's a list of chain names including any new chains added to the local registry
+};
+
+export const environmentCommandOption: Options = {
+  type: 'string',
+  description: 'The name of the environment to deploy to',
+  alias: 'e',
+  choices: ChainTypes,
+};
 
 export const outputFileCommandOption = (
   defaultPath?: string,
