+++ typescript/cli/src/commands/types.ts
@@ -1,0 +1,2 @@
+export const ChainTypes = ['mainnet', 'testnet'];
+export type ChainType = (typeof ChainTypes)[number];
